refactor: modify ImpactCalculationStrategy interface and its implementations to use double precision for the calculation of impact, and updated ProductServiceTest class accordingly.

+++ app/src/main/java/com/mightyfour/application/ProductService.java
@@ -23,7 +23,7 @@ public class ProductService {
     }
 
     public ProvideImpactValueResult calculateImpact(UUID productId, String strategyNum){
-        int result = (factory.create(strategyNum)).calculateImpact(repo.findProduct(productId));
+        double result = (factory.create(strategyNum)).calculateImpact(repo.findProduct(productId));
         return new ProvideImpactValueResult(result);
     }
